+++ src/components/spartan/STab/variants/underline/UnderlineTab.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts"></script>
+
+<template>
+    <nav class="border-b border-gray-200" aria-label="Tabs">
+        <ul class="-mb-px flex space-x-8">
+            <slot />
+        </ul>
+    </nav>
+</template>