+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- `STabs` components implementation with `underline`, `pills` and `vetches` variants.
 
 ## 2.0.0-beta.26 - 2023-10-27